* Fixed various PR issues * Added icon for Win32 client * Enabled UI button tab navigation for gamepad * Fixed a couple of incorrect gamepad input mappings * Mitigated LocalPlayer::isImmobile() bug to gamepad only * Tweaked ControllerTurnInput to closer match Legacy Console Edition's turn behavior * Documented various fields * Made timing more percise for Windows

+++ source/client/player/input/UnifiedTurnBuild.cpp
@@ -56,7 +56,7 @@ void UnifiedTurnBuild::setScreenSize(int width, int height)
 
 TurnDelta UnifiedTurnBuild::getTurnDelta()
 {
-	float timeS = getTimeS();
+	double timeS = getTimeS();
 
 	float m1 = 0.0f, m2 = 0.0f;
 	float xd = 0.0f, yd = 0.0f;
